修改整车控制系统存在问题，修改电池管理存在问题

+++ Yamaha/src/CustomCarMsgWorker.hpp
@@ -340,8 +340,8 @@ signals:
     void bmsVersionChanged(float value);
     void bmsModuleAmountChanged(float value);
     void bmsLifeChanged(float value);
-    void batteryPackHighestTempChanged(float value);
-    void batteryPackMinTempChanged(float value);
+    void batteryPackHighestTempChanged(int value);
+    void batteryPackMinTempChanged(int value);
     void batteryManagePackEnergyChanged(float value);
     void bmsControlStatusChanged(float value);
     void batAmountChanged(float value);
